Just keep churning

+++ app/src/main/kotlin/d5700/instruction/SetTInstruction.kt
@@ -10,6 +10,6 @@ class SetTInstruction : Instruction() {
     }
 
     override fun perform(cpu: CPU) {
-        cpu.setTimer(value)
+        cpu.setTimer(value.toByte())
     }
 }
